+++ src/commands/module.js
@@ -1,5 +1,13 @@
+import { arrayify } from '../util/arrayify.js';
 import { expand } from '../util/expand.js';
 import { TiError } from '../util/tierror.js';
+import { existsSync } from 'node:fs';
+import { dirname, join, parse } from 'node:path';
+import * as timodule from '../util/timodule.js';
+import chalk from 'chalk';
+import { capitalize } from '../util/capitalize.js';
+
+const { bold, cyan, gray } = chalk;
 
 // if a platform is not in this map, then we just print the capitalized platform name
 const platformNames = {
@@ -78,7 +86,7 @@ ModuleSubcommands.list = {
 	async fn(logger, config, cli) {
 		const isJson = cli.argv.json || cli.argv.output === 'json';
 		const projectDir = cli.argv['project-dir'];
-		const p = expand(projectDir || '.');
+		let p = expand(projectDir || '.');
 		const searchPaths = {
 			project: [],
 			config: [],
@@ -89,82 +97,84 @@ ModuleSubcommands.list = {
 			config: 'Configured Path Modules',
 			global: 'Global Modules'
 		};
-		const confPaths = config.get('paths.modules');
+		const confPaths = arrayify(config.get('paths.modules'), true);
 		const defaultInstallLocation = cli.env.installPath;
 		const sdkLocations = cli.env.os.sdkPaths.map(p => expand(p));
 
-		console.log(sdkLocations);
-
-		/*
 		// attemp to detect if we're in a project folder by scanning for a tiapp.xml
 		// until we hit the root
-		if (fs.existsSync(p)) {
-			while (path.dirname(p) !== p) {
-				if (fs.existsSync(path.join(p, 'tiapp.xml'))) {
-					fs.existsSync(p = path.join(p, 'modules')) && searchPaths.project.push(p);
+		if (existsSync(p)) {
+			for (const { root } = parse(p); p !== root; p = dirname(p)) {
+				if (existsSync(join(p, 'tiapp.xml'))) {
+					p = join(p, 'modules');
+					if (existsSync(p)) {
+						searchPaths.project.push(p);
+					}
 					break;
 				}
-				p = path.dirname(p);
 			}
 		}
 
 		// set our paths from the config file
-		Array.isArray(confPaths) || (confPaths = [ confPaths ]);
-		confPaths.forEach(function (p) {
-			p && fs.existsSync(p = afs.resolvePath(p)) && searchPaths.project.indexOf(p) === -1 && searchPaths.config.indexOf(p) === -1 && (searchPaths.config.push(p));
-		});
+		for (let path of confPaths) {
+			path = expand(path);
+			if (existsSync(path) && !searchPaths.project.includes(path) && !searchPaths.config.includes(path)) {
+				searchPaths.config.push(p);
+			}
+		}
 
 		// add any modules from various sdk locations
-		sdkLocations.indexOf(defaultInstallLocation) === -1 && sdkLocations.push(defaultInstallLocation);
-		_cli.sdk && sdkLocations.push(afs.resolvePath(_cli.sdk.path, '..', '..', '..'));
-		sdkLocations.forEach(function (p) {
-			fs.existsSync(p = afs.resolvePath(p, 'modules')) && searchPaths.project.indexOf(p) === -1 && searchPaths.config.indexOf(p) === -1 && searchPaths.global.indexOf(p) === -1 && (searchPaths.global.push(p));
-		});
+		if (!sdkLocations.includes(defaultInstallLocation)) {
+			sdkLocations.push(defaultInstallLocation);
+		}
+		if (cli.sdk) {
+			sdkLocations.push(expand(cli.sdk.path, '..', '..', '..'));
+		}
+		for (let path of sdkLocations) {
+			path = expand(path, 'modules');
+			if (existsSync(path) && !searchPaths.project.includes(path) && !searchPaths.config.includes(path) && !searchPaths.global.includes(path)) {
+				searchPaths.global.push(path);
+			}
+		}
 
-		const results = await new Promise(resolve => appc.timodule.scopedDetect(searchPaths, config, null, resolve));
+		const results = await timodule.detect(searchPaths, config, logger);
 
 		if (isJson) {
-			// alias ios to iphone for backwards compatibility
-			if (results.global && results.global.ios) {
-				results.global.iphone = results.global.ios;
-			}
 			console.log(JSON.stringify(results, null, '\t'));
-		} else {
-			logger.banner();
-
-			Object.keys(results).forEach(function (scope) {
-				var modules = results[scope],
-					platforms = Object.keys(modules);
-
-				if (scope === 'project' && projectDir === undefined && !platforms.length) {
-					// no sense printing project modules if there aren't any and the
-					// user never asked to see them
-					return;
-				}
-
-				_logger.log(scopeLabels[scope].bold);
-
-				if (platforms.length) {
-					platforms.forEach(function (platform, i) {
-						i && _logger.log(); // add some whitespace
-
-						var platformName = platformNames[platform.toLowerCase()] || appc.string.capitalize(platform);
-						_logger.log(platformName.grey);
-
-						Object.keys(modules[platform]).forEach(function (name) {
-							_logger.log('  ' + name);
-							Object.keys(modules[platform][name]).forEach(function (ver) {
-								_logger.log('    ' + appc.string.rpad(ver, 7).cyan + ' ' + modules[platform][name][ver].modulePath);
-							});
-						});
-					});
-				} else {
-					_logger.log(__('No modules found').grey);
-				}
-
-				_logger.log();
-			});
+			return;
 		}
-		*/
+
+		logger.skipBanner(false);
+		logger.banner();
+
+		for (const [scope, modules] of Object.entries(results)) {
+			const platforms = Object.keys(modules);
+			if (scope === 'project' && projectDir === undefined && !platforms.length) {
+				// no sense printing project modules if there aren't any and the
+				// user never asked to see them
+				continue;
+			}
+
+			console.log(bold(scopeLabels[scope]));
+			if (platforms.length) {
+				let i = 0;
+				for (const platform of platforms) {
+					if (i++) {
+						console.log();
+					}
+
+					const platformName = platformNames[platform.toLowerCase()] || capitalize(platform);
+					console.log(gray(platformName));
+					for (const [name, versions] of Object.entries(modules[platform])) {
+						console.log(`  ${name}`);
+						for (const [ver, mod] of Object.entries(versions)) {
+							console.log(`    ${cyan(ver.padEnd(7))} ${mod.modulePath}`);
+						}
+					}
+				}
+			} else {
+				console.log(gray('No modules found'));
+			}
+		}
 	}
 }